Extract OpenBSD/NetBSD CentralProcessor common bases (BSD dedup)
The OpenBSD JNA and FFM CentralProcessors were identical except for the
sysctl backend and the native CPU-tick/load-average calls, and the
dmesg-based topology/cache parsing was duplicated again in NetBSD.

Add a cross-BSD BsdCentralProcessor base (oshi.hardware.common.platform.unix)
holding the shared dmesg processor/cache enumeration, vmstat-backed context
switch/interrupt counters, and an abstract int sysctl(name, def) hook plus an
abstract queryVmStats(). NetBsdCentralProcessor and a new OpenBsdCentralProcessor
extend it; OpenBsdCentralProcessor adds the shared processor identification,
frequency, and OpenBSD vmstat parsing with abstract sysctl hooks, leaving only
the native tick readers and getloadavg in the JNA and FFM subclasses.

This fixes a pre-existing NetBSD bug: its initProcessorCounts split the dmesg
cache line on regex group(1) (the cpu number) instead of group(2) (the cache
description), so dmesg-derived cache detection never produced any caches; the
shared base uses the correct group(2).
